pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "worktime-tracker"
version = "0.1.0"
description = "Flask + SQLite app per tracking ore, multi-utente, con mini dashboard."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "flask>=3.0",
  "flask-sqlalchemy>=3.1",
  "sqlalchemy>=2.0",
  "flask-login>=0.6",
  "flask-wtf>=1.2",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app_flask", "tests"]
exclude = [
  "migrations/*",
  "infra/*",
  "docs/*",
]

[tool.ruff.lint]
select = [
  "E", "F",         # pyflakes/pycodestyle base
  "I",              # import sort
  "UP",             # pyupgrade
  "B",              # bugbear
  "ANN",            # type annotations
  "C90",            # mccabe complexity
]
ignore = [
  "ANN101", "ANN102",     # niente typing obbligatorio per self/cls
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q --strict-markers --maxfail=1 --disable-warnings"
markers = [
  "slow: test lenti/integrati",
]
pythonpath = [
  ".", "app_flask"
]
filterwarnings = [
  "error::DeprecationWarning:app_flask.*",
  "ignore::DeprecationWarning:sqlalchemy.*",
]

[tool.coverage.run]
branch = true
source = ["app_flask"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 85
skip_covered = true

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
no_implicit_optional = true
strict_optional = true
exclude = "(migrations/|tests/|infra/)"

[tool.isort]
profile = "black"
